pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finops-cost-capitalization-pipeline"
version = "0.3.0"
description = "Portfolio-grade Analytics Engineering project for FinOps OPEX vs CAPEX recommendation workflows."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Bruno Ramos Martins" }
]
dependencies = [
  "pandas>=2.2,<3.0",
  "numpy>=2.0,<3.0",
  "faker>=25.0,<30.0",
  "pyarrow>=16.0,<20.0",
  "pyyaml>=6.0,<7.0",
  "duckdb>=1.1,<2.0",
  "dbt-core>=1.8,<2.0",
  "dbt-duckdb>=1.8,<2.0",
  "dagster>=1.8,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5,<0.8",
  "pre-commit>=3.7,<5.0",
  "sqlfluff>=3.1,<4.0",
  "sqlfluff-templater-dbt>=3.1,<4.0",
]

[project.scripts]
finops-generate = "finops_capex.cli.generate_billing_data:main"
finops-run-pipeline = "finops_capex.pipeline.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["finops_capex"]
